+++ src/main/java/roomescape/controller/ReservationTimeController.java
@@ -4,7 +4,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 import roomescape.dao.ReservationTimeDAO;
-import roomescape.model.ReservationTime;
+import roomescape.domain.ReservationTime;
 
 import java.util.List;
